requirements.txt
astroid==2.3.3
autopep8==1.5
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.761
mypy-extensions==0.4.3
pkg-resources==0.0.0
pycodestyle==2.5.0
pylint==2.4.4
PySDL2==0.9.7
rope==0.16.0
six==1.14.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
wrapt==1.11.2
